Add Loggly test
Add logging to Loggly with winston and morgan based on info from:

* https://www.loggly.com/docs/nodejs-logs/
* http://tostring.it/2014/06/23/advanced-logging-with-nodejs/
* https://github.com/imperugo/NodeJs-Sample/blob/master/Logging/WinstonSample/express-configuration/index.js
* https://github.com/expressjs/morgan

For this example, all requests will be logged in standard Apache
combined log both to the console and to Loggly. In order to run this
version of the app you must provide the correct Loggly token in the
LOGGLY_TOKEN environment variable.